\name{estimator-class}
\docType{class}
\alias{estimator-class}
\title{Class \code{"estimator"}}
\description{
Class for storing java Estimator object.
}
\section{Slots}{
  \describe{
    \item{\code{java}:}{Object of class \code{"jobjRef"} ~~ }
  }
}

\author{Kostin Andrey <andrey.kostin@portfolioeffect.com>}
\examples{
showClass("estimator")
}
\keyword{PortfolioEffectEstim,nonparametric, models}
%%\concept{high frequency, market data, optimization, realtime analytics, risk, risk management, trading, trading strategies}
\keyword{classes}